activation manager and survey into extension top
