+++ src/webviews/extension-side/serviceRegistry.node.ts
@@ -2,7 +2,11 @@
 // Licensed under the MIT License.
 'use strict';
 
-import { IExtensionSingleActivationService, IExtensionSyncActivationService } from '../../platform/activation/types';
+import {
+    IExtensionActivationManager,
+    IExtensionSingleActivationService,
+    IExtensionSyncActivationService
+} from '../../platform/activation/types';
 import { IWebviewViewProvider, IWebviewPanelProvider } from '../../platform/common/application/types';
 import { IServiceManager } from '../../platform/ioc/types';
 import { INotebookWatcher, IVariableViewProvider } from './variablesView/types';
@@ -42,6 +46,8 @@ import { AmlComputeContext } from './amlContext.node';
 import { IImportTracker, ImportTracker } from './importTracker.node';
 import { GlobalActivation } from './globalActivation';
 import { WorkspaceActivation } from './workspaceActivation.node';
+import { ExtensionActivationManager } from './activationManager';
+import { DataScienceSurveyBanner, ISurveyBanner } from './survey/dataScienceSurveyBanner.node';
 
 export function registerTypes(serviceManager: IServiceManager, _isDevMode: boolean) {
     serviceManager.addSingleton<IExtensionSingleActivationService>(IExtensionSingleActivationService, GlobalActivation);
@@ -110,4 +116,9 @@ export function registerTypes(serviceManager: IServiceManager, _isDevMode: boole
     );
 
     serviceManager.addSingletonInstance<IExtensionSideRenderer>(IExtensionSideRenderer, new ExtensionSideRenderer());
+
+    serviceManager.addSingleton<ISurveyBanner>(ISurveyBanner, DataScienceSurveyBanner);
+    serviceManager.addBinding(ISurveyBanner, IExtensionSingleActivationService);
+    // Activation Manager
+    serviceManager.add<IExtensionActivationManager>(IExtensionActivationManager, ExtensionActivationManager);
 }
